refactor(landing): add landing page with logo, register & join as guest buttons
